requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
flake8==3.8.4
gunicorn==20.0.4
Markdown==3.3.3
mccabe==0.6.1
p==0.0.0
Pillow==8.0.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
reportlab==3.5.55
sorl-thumbnail==12.6.3
sqlparse==0.4.1
toml==0.10.1
